Bugfix: Fix duplicate caching of Rollout and ActionStatus
Due to the use of @CacheField within JpaActionStatus there will be a cache created named "JpaActionStatus" (see implementation CacheFieldEntityListener.java).
The class "CacheWriteNotify" try to access the same cache as well but use interface of JpaActionStatus instead. This results in a new, different cache. So in fact we are caching it twice.

Same was also happening to Rollout. By using JPA entity's implementation instead of the interface both calls to cache manager will return the same cache now.

Signed-off-by: Dominic Schabel <dominic.schabel@bosch-si.com>