New service which is responsible for execution of TargetFiterQueries
The intention of this change is to have one place where external search engines can be plugged into hawkbit (like other device-management systems or also search-related DBs like elastic search or Solr.

Next steps will be to enable different kind of TargetFilterQuery implementations which will enable to leverage the extended capabilities of the different search engines up to the UI and the user.

Signed-off-by: Peter Vigier <Peter.Vigier@bosch-si.com>