Add springdoc-openapi to management rest API
This enables a swagger-ui interface to test and discover the API.
And it also notably generates an OpenAPI specification file which could
then be used to generate API clients in multiple languages to interact
with hawkBit.

Compared to https://github.com/kokuwaio/hawkbit-specs , this approach
has the advantage that the specification file can be easily updated
whenever the API is updated.
Also the generated specification is complete.

A similar approach could be used to generate the OpenAPI specification
for the other APIs as asked
[here](https://gitter.im/eclipse/hawkbit?at=5e1c2db3b720fa5b3c06a101).

Signed-off-by: Vladimir Svoboda <vsvoboda@forkoder.eu>