hawkbit-ddi-resource: do not log range requests
HawkBit creates ActionStatus entries for each GET request associated with an
artefact of an assigned software module. Amongst others, this severely breaks
RAUC block-based adaptive updates [1], which issue a flurry of range requests
for small chunks, so that the whole image does not have to be streamed in case
of incremental system changes.

[1] https://rauc.readthedocs.io/en/latest/advanced.html#block-based-adaptive-update-block-hash-index

Fixes: https://github.com/eclipse/hawkbit/issues/1249

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>