Fix handling of artifact range requests
Artifact range requests in HawkBit were impossible to do efficiently, as the
underlying mechanism had no choice but to read the start byte and discard
content both before and after.

To address this situation, we need to introduce a new method to DbArtifact:
getFileInputStream(long, long), which maps to the semantics of a range request.

The next step is to adjust FileStreamingUtil to use it for both simple range
and multi-part requests.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>