Fix MgmtBaseEntity deserialization (#1633)
Before REST refactoring MgmtBaseEntity was able to deserialize fields
like createdBy. After refactoring, with READ_ONLY access it was dropped
and these fields become null. While this could be a good change, it is
not backward compatible (is that needed) and most importantly will lead
to the fact that Feign client won't be able to access that
data. So, at least for now, I return deserialization back

Signed-off-by: Marinov Avgustin <Avgustin.Marinov@bosch.com>