simple-ui: Also show target attributes in Target's view (#2304)
It is useful to see target attributes in the Target's view: for instance
SWUpdate can deliver values from the platform which identify some
features of it. This can include an arbitrary data, e.g. platform HW
revision, component's versions etc.
Thus, add a possibility to see those attributes in the dedicated view.

Signed-off-by: Oleksandr Andrushchenko <andr2000@gmail.com>